+++ app/src/main/java/com/example/resona/data/remote/model/CategoryRequest.kt
@@ -1,4 +1,4 @@
-package com.example.resona.data.dto
+package com.example.resona.data.remote.model
 
 data class CategoryRequest(
     val category: String,
