+++ app/src/main/java/com/example/resona/data/remote/model/ProfileImageResponse.kt
@@ -1,4 +1,4 @@
-package com.example.resona.data.dto
+package com.example.resona.data.remote.model
 
 data class ProfileImageResponse (
     val profileImage : String
