+++ app/src/main/java/com/example/resona/data/remote/model/ProfileImageViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.resona.data.dto.ProfileImageResponse
 import com.example.resona.data.repository.MainRepository
 import com.example.resona.data.utils.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -35,7 +34,9 @@ class ProfileImageViewModel @Inject constructor(
                 is ApiResult.Error -> {
                     _error.value = result.exception.message
                 }
+                is ApiResult.Loading -> {
+                }
+                }
             }
-        }
     }
 }
